+++ src/backend/src/core/setting.py
@@ -1,6 +1,11 @@
 import os
 from typing import Final
 
+from dotenv import load_dotenv
+
+
+load_dotenv()
+
 
 class _NoArg:
     pass
@@ -16,3 +21,6 @@ def get_env_var(key: str, default: str | _NoArg = NO_ARG) -> str:
         if isinstance(default, _NoArg):
             raise ValueError(f"Environment variable {key} is missing") from exception
         return default
+
+
+DATABASE_URL: Final[str] = get_env_var("DATABASE_URL")